refactor: collapse npm barrel into npm/index.ts
